Creating User model.
