+++ kotlin-project/app/src/main/java/com/eps/creditoffer/Connections/StartLink.kt
@@ -8,7 +8,6 @@ import java.util.*
 
 class StartLink {
 
-    // private val ip: String = "192.168.0.16"
     private val ip: String = "10.0.2.2"
 
     fun start(nome: String): Boolean {
