+++ kotlin-project/app/src/main/java/com/eps/creditoffer/Connections/UserLink.kt
@@ -1,5 +1,6 @@
 package com.eps.creditoffer.Connections
 
+import com.eps.creditoffer.Models.User
 import com.github.kittinunf.fuel.Fuel
 import com.github.kittinunf.fuel.core.ResponseDeserializable
 import com.github.kittinunf.fuel.core.extensions.jsonBody
@@ -9,94 +10,86 @@ import org.json.JSONObject
 
 class UserLink {
 
-    var cpf: Int = 0
-    var name: String = ""
-    var email: String = ""
-    var phone: Int = 0
-    var monthlyIncome: Float = 0F
-    var debt: List<OverdraftDebtLink> = emptyList()
+    companion object {
+        private val ip: String = "10.0.2.2"
 
-    // private val ip: String = "192.168.0.16"
-    private val ip: String = "10.0.2.2"
+        fun get(id: Int): Boolean {
+            println("----UserLink.get----")
+            val url: String = "http://" + ip + ":3000/api/users/" + id.toString()
 
-    class Deserializer : ResponseDeserializable<UserLink> {
-        override fun deserialize(content: String) = Gson().fromJson(content, UserLink::class.java)
-    }
-
-    fun get(id: Int): Boolean {
-        println("----UserLink.get----")
-        val url: String = "http://" + ip + ":3000/api/users/" + id.toString()
-
-        val (request, response, result) = Fuel.get(url)
-            .responseObject(Deserializer())
-        println("Response:" + response)
-        val (bytes, error) = result
-        if (bytes != null) {
-            this.cpf = bytes.cpf
-            this.name = bytes.name
-            this.email = bytes.email
-            this.phone = bytes.phone
-            this.monthlyIncome = bytes.monthlyIncome
-        }
-        when (result) {
-            is Result.Success -> {
-                println("Success")
-                return true
+            val (request, response, result) = Fuel.get(url)
+                .responseObject(User.Deserializer())
+            println("Response:" + response)
+            val (bytes, error) = result
+            if (bytes != null) {
+                User.id = bytes.id
+                User.cpf = bytes.cpf
+                User.name = bytes.name
+                User.email = bytes.email
+                User.phone = bytes.phone
+                User.monthlyIncome = bytes.monthlyIncome
             }
-            is Result.Failure -> {
-                println("Failure")
-                return false
+            when (result) {
+                is Result.Success -> {
+                    println("Success")
+                    return true
+                }
+                is Result.Failure -> {
+                    println("Failure")
+                    return false
+                }
             }
         }
-    }
 
-    fun create() {
-        println("----UserLink.create----")
-        val url: String = "http://" + ip + ":3000/api/users"
+        fun create() {
+            println("----UserLink.create----")
+            val url: String = "http://" + ip + ":3000/api/users"
 
-        val json = JSONObject()
-        json.put("cpf", this.cpf)
+            val json = JSONObject()
+            json.put("cpf", User.cpf)
 
-        val (request, response, result) = Fuel.post(url)
-            .jsonBody(json.toString())
-            .response()
-        println("Response:" + response)
-        val (bytes, error) = result
-        if (bytes != null) {
-        }
-        println(result)
-        when (result) {
-            is Result.Success -> {
-                println("Success")
+            val (request, response, result) = Fuel.post(url)
+                .jsonBody(json.toString())
+                .response()
+            println("Response:" + response)
+            val (bytes, error) = result
+            if (bytes != null) {
             }
-            is Result.Failure -> {
-                println("Failure")
+            println(result)
+            when (result) {
+                is Result.Success -> {
+                    println("Success")
+                }
+                is Result.Failure -> {
+                    println("Failure")
+                }
             }
         }
-    }
 
-    fun listDebt(id: Int): Boolean {
-        println("----User.listDebt---")
+        fun listDebt(id: Int): Boolean {
+            println("----User.listDebt---")
 
-        val url: String = "http://" + ip + ":3000/api/overdraftDebts/" + id.toString() + "/listDebt"
-        val (request, response, result) = Fuel.get(url)
-            .responseObject(OverdraftDebtLink.ListDeserializer())
-        println("Response:" + response)
-        val (bytes, error) = result
-        print("Bytes: " + bytes)
+            val url: String = "http://" + ip + ":3000/api/overdraftDebts/" + id.toString() + "/listDebt"
+            val (request, response, result) = Fuel.get(url)
+                .responseObject(OverdraftDebtLink.ListDeserializer())
+            println("Response:" + response)
+            val (bytes, error) = result
+            print("Bytes: " + bytes)
 
-        if (bytes != null) {
-            this.debt = bytes
-        }
-        when (result) {
-            is Result.Success -> {
-                println("Success")
-                return true
+            if (bytes != null) {
+                User.debt = bytes
             }
-            is Result.Failure -> {
-                println("Failure")
-                return false
+            when (result) {
+                is Result.Success -> {
+                    println("Success")
+                    return true
+                }
+                is Result.Failure -> {
+                    println("Failure")
+                    return false
+                }
             }
         }
     }
+
 }
